+++ packages/amplify-category-api/src/commands/api/override.ts
@@ -7,8 +7,8 @@ import {
   generateOverrideSkeleton,
   pathManager,
   stateManager,
-} from 'amplify-cli-core';
-import { printer, prompter } from 'amplify-prompts';
+} from '@aws-amplify/amplify-cli-core';
+import { printer, prompter } from '@aws-amplify/amplify-prompts';
 import * as path from 'path';
 import { ADMIN_QUERIES_NAME } from '../../category-constants';
 import { AdminQueriesProps, ApigwInputState } from '../../provider-utils/awscloudformation/apigw-input-state';