+++ packages/amplify-category-api/src/commands/api/remove.ts
@@ -1,5 +1,5 @@
-import { $TSContext, AmplifyCategories, AmplifySupportedService } from 'amplify-cli-core';
-import { printer } from 'amplify-prompts';
+import { $TSContext, AmplifyCategories, AmplifySupportedService } from '@aws-amplify/amplify-cli-core';
+import { printer } from '@aws-amplify/amplify-prompts';
 import * as path from 'path';
 
 const subcommand = 'remove';