+++ packages/amplify-graphql-auth-transformer/src/__tests__/field-auth-argument.test.ts
@@ -5,7 +5,7 @@ import { AppSyncAuthConfiguration } from '@aws-amplify/graphql-transformer-inter
 import { AuthTransformer } from '../graphql-auth-transformer';
 import { featureFlags } from './test-helpers';
 
-jest.mock('amplify-prompts');
+jest.mock('@aws-amplify/amplify-prompts');
 
 test('subscriptions are only generated if the respective mutation operation exists', () => {
   const validSchema = `