+++ packages/amplify-schema-validator/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.2](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-schema-validation@0.2.1...@aws-amplify/graphql-schema-validation@0.2.2) (2023-03-30)
+
+**Note:** Version bump only for package @aws-amplify/graphql-schema-validation
+
 ## [0.2.1](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-schema-validation@0.2.0...@aws-amplify/graphql-schema-validation@0.2.1) (2023-03-15)
 
 **Note:** Version bump only for package @aws-amplify/graphql-schema-validation